Refactor Mock Recording and Playback into mock.js
