+++ app.js
@@ -2,25 +2,24 @@
 // Read and Write. (Save Gnip costs and ArcGIS Online writes).
 // Write without Reading. (Save Gnip costs, but still write to ArcGIS Online).
 // Read without Writing? <-- Not yet.
-var mock = process.env.MOCK === undefined?false:process.env.MOCK,
-    mockRead = mock || (process.env.MOCKREAD === undefined?false:process.env.MOCKREAD),
-    mockWrite = mock || (process.env.MOCKWRITE === undefined?false:process.env.MOCKWRITE);
+var mock = require('./lib/mock');
 
-if (mockWrite && mockRead && mockWrite !== mockRead) {
+if (mock.mockWrite && mock.mockRead && mock.mockWrite !== mock.mockRead) {
   console.error('If you specify a MOCKREAD and MOCKWRITE setting, they must be the same.');
   process.exit(1); 
 }
 
-if (mockWrite && !mockRead) {
+if (mock.mockWrite && !mock.mockRead) {
   console.error('You cannot currently Mock Writing whilst reading real Gnip data');
   process.exit(1);
 }
 
-var mocking = mock || mockRead || mockWrite,
-    mockString = mocking?('Mock [' + 
-                (mockRead===false?'':('Read' + (mockWrite?'/Write ':' ') + mockRead)) + 
+var mockString = mock.mocking?('Mock [' + 
+                (mock.mockRead===false?'':('Read' + (mock.mockWrite?'/Write ':' ') + mock.mockRead)) + 
                 ']'):'';
 
+
+// And on to the real meat of the project.
 var env = process.env.NODE_ENV = process.env.NODE_ENV || 'development';
 var clientId = process.env.CLIENT_ID = process.env.CLIENT_ID || '';
 
@@ -75,6 +74,6 @@ app.all (paths.queryStatus, handlers.handleQueryStatus);
 app.listen(process.env.PORT || 1337,  function() {
   console.log('Listening at http://%s:%d/', this.address().address, this.address().port);
   console.log('Env: ' + env);
-  console.log(mocking?mockString:'Mock: N/A');
+  console.log(mock.mocking?mockString:'Mock: N/A');
   console.log('Client ID: ' + clientId);
 });
