+++ lib/handlers.js
@@ -6,21 +6,13 @@ var uuid = require('node-uuid'),
 var writers = require('./writers'),
     readers = require('./readers'),
     querySession = require('./query-session'),
-    wq = require('./write-queue');
+    wq = require('./write-queue'),
+    mock = require('./mock');
 
 var Portal = require('./portal');
 
 var qsStatus = require('./query-session-status'),
     pageStatus = require('./page-status');
-
-var env = process.env.NODE_ENV = process.env.NODE_ENV || 'development';
-var mock = process.env.MOCK === undefined?false:process.env.MOCK,
-    mockRead = mock || (process.env.MOCKREAD === undefined?false:process.env.MOCKREAD),
-    mockWrite = mock || (process.env.MOCKWRITE === undefined?false:process.env.MOCKWRITE);
-var recordOutput = process.env.RECORDGNIP = ((process.env.RECORDGNIP || false) === 'true');
-
-var outputRoot = './output/',
-    mockRoot = './mocks/';
 
 function sendError(res, err) {
   var noLog,
@@ -169,16 +161,14 @@ function handlePushQuery(req, res) {
         // Got a page of Gnip records.
         qs.readProgress = estimatedProgress;
 
-        if (!mockRead) {
-          recordOutputToFile('GnipPage', qs, {
-            gnipRecords: gnipRecords, 
-            pageNumber: pageNumber, 
-            estimatedProgress: estimatedProgress
-          });
-        }
+        mock.recordMockOutput('GnipPage', qs, { // Will only write if RECORDGNIP=true
+          gnipRecords: gnipRecords, 
+          pageNumber: pageNumber, 
+          estimatedProgress: estimatedProgress
+        });
 
-        if (mockWrite) {
-          return true;
+        if (mock.mockWrite) {
+          return true; // Let's not actually write anything.
         }
 
         // Break down the page of Gnip records into chunks small enough to write.
@@ -242,107 +232,8 @@ function handlePushQuery(req, res) {
   });
 }
 
-function recordOutputToFile(writeType, qs, output) {
-  if (recordOutput == true) {
-    if ((writeType === 'GnipPage' && !mockRead) ||
-        (writeType === 'QueryStatus' && !mockWrite)) {
-      // We'll log the output to a folder and files within that folder.
-      // Can use some of these files later for mock mode if need be.
-      var countProperty = '_dev_count_' + writeType;
-      if (qs.hasOwnProperty(countProperty)) {
-        qs[countProperty] += 1;
-      } else {
-        qs[countProperty] = 0;
-      }
-
-      var outputPage = qs[countProperty];
-
-      if (writeType === 'GnipPage') {
-        outputPage = output.pageNumber;
-      }
-
-      var fs = require('fs'),
-          mkdirp = require('mkdirp'),
-          folder = outputRoot + qs.id + '/' + writeType,
-          filename = folder + '/' + outputPage + '.json';
-
-      if (!fs.existsSync(folder)) {
-        mkdirp.sync(folder);
-      }
-
-      fs.writeFile(filename, JSON.stringify(output, null, "  "), function(err) {
-        if (!err) {
-          console.log('Wrote output to ' + filename);
-          // qs.writtenOutputToFile = true;
-        } else {
-          console.error('Failed to write output to ' + filename);
-          console.error(err);
-          // qs.writtenOutputToFile = false;
-        }
-      });
-    }
-  }
-}
-
-var mockStart = null,
-    mockFiles = [],
-    finalResponse = null;
-function handleMockQueryStatus(req, res) {
-  if (mockWrite) {
-    var fs = require('fs');
-
-    if (mockStart === null) {
-      // Initialize…
-      function setupFilenames(dirPath, fileRegex) {
-        var allFilenames = require('fs').readdirSync(dirPath),
-            result = [];
-        for (var i=0; i<allFilenames.length; i++) {
-          var filename = allFilenames[i];
-          if (fileRegex.test(filename)) {
-            result.push(dirPath + '/' + filename);
-          }
-        }
-        return result;
-      }
-
-      mockStart = new Date();
-      finalResponse = { error: 'Something went wrong reading mock data - please report to the server administrator.' };
-      mockFiles = setupFilenames(mockRoot + mock + '/QueryStatus', /^[0-9]+.json/);
-    }
-
-    var currentMock;
-    if (mockFiles.length > 0) {
-      currentMock = JSON.parse(fs.readFileSync(mockFiles.shift(), 'utf8'));
-      if (currentMock.status === 'finished') {
-        // That was the last one. We're done here.
-        currentMock.endTime = new Date();
-        finalResponse = currentMock;
-      }
-    } else {
-      // Report the final response again.
-      console.warn('Duplicate 100% response requested!!!');
-      currentMock = finalResponse;
-    }
-
-    currentMock.id = req.query.queryId;
-    currentMock.startTime = mockStart;
-
-    res.status(200).send(currentMock);
-
-    if (currentMock == finalResponse) {
-      setTimeout(function() {
-        mockStart = null;
-        console.log('RESET MOCK: ' + mock);
-      }, 5000);
-    }
-
-    return true;
-  }
-  return false;
-}
-
 function handleQueryStatus(req, res) {
-  if (handleMockQueryStatus(req, res)) {
+  if (mock.handleQueryStatus(req, res)) {
     // We're in mock mode. No need to do any real work.
     return;
   }
@@ -351,7 +242,7 @@ function handleQueryStatus(req, res) {
   var qs = querySession.get(queryId),
       output = querySession.forOutput(qs);
   if (qs !== undefined) {
-    recordOutputToFile('QueryStatus', qs, output);
+    mock.recordMockOutput('QueryStatus', qs, output); // Will only write if RECORDGNIP=true
     res.status(200).send(output);
   } else {
     res.status(400).send('Invalid query sesson ID: ' + queryId);
